Bookings table updated - 1.0.74

+++ database/migrations/2022_07_21_090921_create_booking_users_table.php
@@ -15,8 +15,8 @@ class CreateBookingUsersTable extends Migration
     {
         Schema::create('booking_users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('booking_id');
-            $table->unsignedBigInteger('user_id');
+            $table->integer('booking_id');
+            $table->integer('user_id');
             $table->timestamps();
         });
     }
